requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
Markdown==3.4.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
virtualenv==20.16.4
